[100] implement Camera: Implement photo taking functionality using the camera.
Add the necessary permissions, layouts and logic to take a photo.

Add `file_paths.xml` and `activity_camera_album_test_main.xml`.
Add `CameraAlbumTestMainActivity` to take a photo and show it on `imageView`.
Add `CameraViewModel` to handle the data of bitmap.
Add permissions for camera and external storage in `AndroidManifest.xml`, and add `FileProvider` for sharing files between apps.
Change the app name to `CameraAlbumTest`.
Change the main activity to `CameraAlbumTestMainActivity`.
Handle the camera permission request logic in `CameraAlbumTestMainActivity`.
Create `output_image.jpg` to store the photo in external storage.
Use `MediaStore.ACTION_IMAGE_CAPTURE` to start camera, and send the uri to camera app.
Update `onCameraResult` to get the photo after taking it.
If get the photo, decode it to Bitmap.
Use `ExifInterface` to check the orientation.
Use Matrix to rotate the photo if required.
Set the rotated bitmap to `imageView`.
Handle the `null` bitmap in the viewModel.